refactor: presign 요청 DTO 3개를 공통 ImagePresignRequest로 통합 + keyPrefix 방식 적용 - 동일하던 Band/Member/PerformancePoster PresignRequest 제거 → global/infra/s3/ImagePresignRequest 단일화 - ImagePresignSupport.issue(request, keyPrefix): objectKeyBuilder 람다 제거, {keyPrefix}/{uuid}.{ext} 키 생성을 내부로 통일 - 3개 컨트롤러/서비스가 공통 DTO 사용, openapi.json 갱신(스키마 3→1)
