+++ src/main/kotlin/com/bandage/bandmanager/domain/member/controller/MemberController.kt
@@ -1,7 +1,6 @@
 package com.bandage.bandmanager.domain.member.controller
 
 import com.bandage.bandmanager.domain.member.dto.req.MemberInfoUpdateRequest
-import com.bandage.bandmanager.domain.member.dto.req.MemberProfileImagePresignRequest
 import com.bandage.bandmanager.domain.member.dto.res.MemberInfoResponse
 import com.bandage.bandmanager.domain.member.dto.res.MemberMetricsResponse
 import com.bandage.bandmanager.domain.member.dto.res.MemberSearchItemResponse
@@ -13,6 +12,7 @@ import com.bandage.bandmanager.facade.dto.MemberJoinRequest
 import com.bandage.bandmanager.facade.dto.MemberResponse
 import com.bandage.bandmanager.global.common.constants.PathPrefix.PREFIX
 import com.bandage.bandmanager.global.common.response.ApiResponse
+import com.bandage.bandmanager.global.infra.s3.ImagePresignRequest
 import com.bandage.bandmanager.global.infra.s3.ImagePresignResponse
 import com.bandage.bandmanager.global.security.annotation.CurrentMemberId
 import com.bandage.bandmanager.global.util.CookieUtil
@@ -74,7 +74,7 @@ class MemberController(
         description = "회원 본인 프로필 이미지를 S3에 PUT 업로드하기 위한 presigned URL을 발급합니다. 응답 objectKey 를 회원 정보 수정 시 profileImg 로 전달합니다.",
     )
     fun issueMemberProfileImagePresignedUrl(
-        @Valid @RequestBody request: MemberProfileImagePresignRequest,
+        @Valid @RequestBody request: ImagePresignRequest,
         @CurrentMemberId memberId: Long,
     ): ApiResponse<ImagePresignResponse> = ApiResponse.success(memberService.issueProfileImagePresignedUrl(request, memberId))
